Move choropleth layers into deprecated folder
