+++ src/layers/deprecated/choropleth-layer/choropleth-layer-fragment.glsl
