+++ src/layers/deprecated/choropleth-layer/choropleth-layer.js
