+++ src/layers/deprecated/choropleth-layer/index.js
