+++ src/layers/deprecated/extruded-choropleth-layer-64/extruded-choropleth-layer-fragment.glsl
